fix usuario
